+++ src/js/client.ts
@@ -59,22 +59,26 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
    * @inheritDoc
    */
   public eventFromMessage(message: string, level?: SeverityLevel, hint?: EventHint): PromiseLike<Event> {
-    return eventFromMessage(this._options.stackParser, message, level, hint, this._options.attachStacktrace).then(
-      (event: Event) => {
-        // TMP! Remove this function once JS SDK uses threads for messages
-        if (!event.exception?.values || event.exception.values.length <= 0) {
+    if (this._options.useThreadsForMessageStack) {
+      return eventFromMessage(this._options.stackParser, message, level, hint, this._options.attachStacktrace).then(
+        (event: Event) => {
+          // TMP! Remove this function once JS SDK uses threads for messages
+          if (!event.exception?.values || event.exception.values.length <= 0) {
+            return event;
+          }
+          const values = event.exception.values.map(
+            (exception: Exception): Thread => ({
+              stacktrace: exception.stacktrace,
+            }),
+          );
+          (event as { threads?: { values: Thread[] } }).threads = { values };
+          delete event.exception;
           return event;
-        }
-        const values = event.exception.values.map(
-          (exception: Exception): Thread => ({
-            stacktrace: exception.stacktrace,
-          }),
-        );
-        (event as { threads?: { values: Thread[] } }).threads = { values };
-        delete event.exception;
-        return event;
-      },
-    );
+        },
+      );
+    }
+
+    return eventFromMessage(this._options.stackParser, message, level, hint, this._options.attachStacktrace);
   }
 
   /**