+++ android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -203,6 +203,9 @@ public class RNSentryModuleImpl {
                 // SentryAndroid needs an empty string fallback for the dsn.
                 options.setDsn("");
             }
+            if (rnOptions.hasKey("sampleRate")) {
+                options.setSampleRate(rnOptions.getDouble("sampleRate"));
+            }
             if (rnOptions.hasKey("sendClientReports")) {
                 options.setSendClientReports(rnOptions.getBoolean("sendClientReports"));
             }
@@ -438,7 +441,7 @@ public class RNSentryModuleImpl {
     }
 
     public void captureReplay(boolean isHardCrash, Promise promise) {
-        Sentry.getCurrentHub().getOptions().getReplayController().sendReplay(isHardCrash, null, null);
+        Sentry.getCurrentHub().getOptions().getReplayController().captureReplay(isHardCrash);
         promise.resolve(getCurrentReplayId());
     }